+++ tasks/romanova_v_jacobi_method/tests/performance/main.cpp
@@ -51,7 +51,7 @@ class RomanovaVJacobiMethodPerfTestProcesses : public ppc::util::BaseRunPerfTest
     if (output_data.size() != exp_answer_.size()) {
       return false;
     }
-    for (int i = 0; i < output_data.size(); i++) {
+    for (size_t i = 0; i < output_data.size(); i++) {
       if (abs(output_data[i] - exp_answer_[i]) > eps_) {
         return false;
       }
